requirements.txt
requests
python-docx
pytest